backend/requirements.txt
fastapi
uvicorn[standard]
kiteconnect
python-dotenv